[rogue] introduce a new generic rotation, fixing all posted rogue crashes; the former is still kept to support multi-target (>=3) situations, but should probably be removed soon [rogue] add support for 2t10 bonus for tricks of the trade, and a similar check for blood elfs' "arcane torrent" ability
