+++ sim/rogue/TestSubtlety.results
@@ -74,8 +74,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-Bandit'sInsignia-40371"
  value: {
-  dps: 8294.27788
-  tps: 5886.09208
+  dps: 8302.07578
+  tps: 5892.45012
  }
 }
 dps_results: {
@@ -125,8 +125,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-BonescytheBattlegear"
  value: {
-  dps: 7128.20117
-  tps: 5059.38962
+  dps: 7128.06098
+  tps: 5059.29009
  }
 }
 dps_results: {
@@ -175,8 +175,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-DarkmoonCard:Death-42990"
  value: {
-  dps: 8245.43389
-  tps: 5851.88277
+  dps: 8242.12071
+  tps: 5849.53042
  }
 }
 dps_results: {
@@ -196,8 +196,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-DeathKnight'sAnguish-38212"
  value: {
-  dps: 8157.1724
-  tps: 5789.07132
+  dps: 8167.97127
+  tps: 5795.9133
  }
 }
 dps_results: {
@@ -231,15 +231,15 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-DislodgedForeignObject-50348"
  value: {
-  dps: 8259.93796
-  tps: 5863.38533
+  dps: 8250.90669
+  tps: 5856.60201
  }
 }
 dps_results: {
  key: "TestSubtlety-AllItems-DislodgedForeignObject-50353"
  value: {
-  dps: 8237.14712
-  tps: 5848.37445
+  dps: 8235.6446
+  tps: 5847.30766
  }
 }
 dps_results: {
@@ -273,8 +273,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-EphemeralSnowflake-50260"
  value: {
-  dps: 8173.9795
-  tps: 5802.75378
+  dps: 8179.37337
+  tps: 5806.58343
  }
 }
 dps_results: {
@@ -322,8 +322,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-ForgeEmber-37660"
  value: {
-  dps: 8166.40265
-  tps: 5794.36136
+  dps: 8161.72382
+  tps: 5791.0394
  }
 }
 dps_results: {
@@ -378,8 +378,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-GnomishLightningGenerator-41121"
  value: {
-  dps: 8237.10224
-  tps: 5845.90324
+  dps: 8242.81008
+  tps: 5849.95581
  }
 }
 dps_results: {
@@ -568,8 +568,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-Shadowblade'sBattlegear"
  value: {
-  dps: 7901.07659
-  tps: 5607.30215
+  dps: 7903.97197
+  tps: 5608.60618
  }
 }
 dps_results: {
@@ -589,8 +589,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-Slayer'sArmor"
  value: {
-  dps: 5490.51321
-  tps: 3895.42103
+  dps: 5488.73564
+  tps: 3893.33952
  }
 }
 dps_results: {
@@ -638,8 +638,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-StormshroudArmor"
  value: {
-  dps: 6171.04692
-  tps: 4379.43616
+  dps: 6162.22026
+  tps: 4373.50418
  }
 }
 dps_results: {
@@ -680,8 +680,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-TerrorbladeBattlegear"
  value: {
-  dps: 7606.71412
-  tps: 5397.43663
+  dps: 7603.41353
+  tps: 5394.73697
  }
 }
 dps_results: {
@@ -701,8 +701,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-ThunderingSkyflareDiamond"
  value: {
-  dps: 8357.93928
-  tps: 5929.49372
+  dps: 8347.34064
+  tps: 5921.96868
  }
 }
 dps_results: {
@@ -736,8 +736,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-TomeofArcanePhenomena-36972"
  value: {
-  dps: 8087.7673
-  tps: 5740.34157
+  dps: 8085.50521
+  tps: 5738.73549
  }
 }
 dps_results: {
@@ -764,8 +764,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-AllItems-VanCleef'sBattlegear"
  value: {
-  dps: 7369.3106
-  tps: 5231.2652
+  dps: 7369.36226
+  tps: 5231.30187
  }
 }
 dps_results: {
@@ -778,8 +778,8 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-Average-Default"
  value: {
-  dps: 8508.99372
-  tps: 6039.04633
+  dps: 8509.30304
+  tps: 6039.19515
  }
 }
 dps_results: {
@@ -869,7 +869,7 @@ dps_results: {
 dps_results: {
  key: "TestSubtlety-SwitchInFrontOfTarget-Default"
  value: {
-  dps: 7719.63138
-  tps: 5478.60469
+  dps: 7720.53051
+  tps: 5478.42353
  }
 }
