+++ sim/rogue/talents.go
@@ -543,14 +543,14 @@ func (rogue *Rogue) registerAdrenalineRushCD() {
 		OnGain: func(aura *core.Aura, sim *core.Simulation) {
 			rogue.ResetEnergyTick(sim)
 			rogue.ApplyEnergyTickMultiplier(1.0)
-			if r, ok := rogue.rotation.(*rotation_generic); ok {
+			if r, ok := rogue.rotation.(*rotation_multi); ok {
 				r.planRotation(sim, rogue)
 			}
 		},
 		OnExpire: func(aura *core.Aura, sim *core.Simulation) {
 			rogue.ResetEnergyTick(sim)
 			rogue.ApplyEnergyTickMultiplier(-1.0)
-			if r, ok := rogue.rotation.(*rotation_generic); ok {
+			if r, ok := rogue.rotation.(*rotation_multi); ok {
 				r.planRotation(sim, rogue)
 			}
 		},
